seperate jsx markup and syles into 2 different files and change js  extensions to jsx
