+++ components/Banner/Banner.jsx
@@ -1,0 +1,14 @@
+import { Wrapper, Container, BannerText } from "./styles";
+
+const Banner = () => {
+    return (
+        <Wrapper>
+            <Container>
+                <BannerText>Customer Satisfaction is our priority</BannerText>
+            </Container>
+        </Wrapper>
+    );
+}
+ 
+export default Banner;
+
