--- components/Footer.js
@@ -1,7 +1,0 @@
-const Footer = () => {
-    return (
-        <div>Footer</div>
-    );
-}
-
-export default Footer;
