+++ components/Footer/styles.jsx
@@ -1,0 +1,30 @@
+import styled from "styled-components";
+
+export const Wrapper = styled.div`
+    margin-top: 50px;
+`
+
+export const Container = styled.div`
+    width: 80%;
+    margin: auto;
+`
+
+export const Title = styled.h3``
+
+export const SubText = styled.p``
+
+export const MainContent = styled.div`
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+`
+
+export const ContactForm = styled.div``
+
+export const MapContent = styled.div`
+    position: relative;
+
+    img{
+        height: 100%;
+        width: 100%;
+    }
+`
