+++ components/Layout.jsx
@@ -1,5 +1,5 @@
-import Footer from "./Footer"
-import Navbar from "./Navbar"
+import Footer from "./Footer/Footer"
+import Navbar from "./Navbar/Navbar"
 
 const Layout = ({children}) => {
     return (
