+++ styles/globals.css
@@ -8,7 +8,10 @@
 	--purple: #764ba2;
 	--blue: #667eea;
 	--grey: #f6f6f6;
-	--gradient: linear-gradient(to bottom, #667eea, #764ba2);
+	--gradient: linear-gradient(to bottom, #667eea, hsl(270, 37%, 46%));
+	--box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23);
+	--box-shadow-hover: 0 14px 28px rgba(0, 0, 0, 0.25),
+		0 10px 10px rgba(0, 0, 0, 0.22);
 }
 
 /* html,
